--- a/instructions.go
+++ b/instructions.go
@@ -54,6 +54,14 @@ func FormatInstructions(b []byte, posOffset int) []string {
 			out = append(out, fmt.Sprintf("%04d %-7s %-5d %-5d",
 				posOffset+i, parser.OpcodeNames[b[i]],
 				operands[0], operands[1]))
+		case 3:
+			out = append(out, fmt.Sprintf("%04d %-7s %-5d %-5d %-5d",
+				posOffset+i, parser.OpcodeNames[b[i]],
+				operands[0], operands[1], operands[2]))
+		case 4:
+			out = append(out, fmt.Sprintf("%04d %-7s %-5d %-5d %-5d %-5d",
+				posOffset+i, parser.OpcodeNames[b[i]],
+				operands[0], operands[1], operands[2], operands[3]))
 		}
 		i += 1 + read
 	}