@geseq 2022-02-19
Hopefully these comments will help with a good starting point to optimize the runtime